add a new svg2pdf.py converter (work in progress)
This converter works like svg2latex.py, extracting all the text from an
SVG file and generating a PDF containing only the graphics, plus a LaTeX
file with all the text (and a reference to the background PDF) embedded
in a picture environment.

However, unlike svg2latex.py, this script constructs its LaTeX file to be
a complete document, and runs pdflatex on it to produce a final PDF
that combines graphics from the SVG, plus all the text, typeset nicely
by LaTeX.

Any text in the svg that has already been rendered and embedded using
the Inkscape textext extension will be extracted and re-typeset as part
of the picture environment, so that it exists as text in the final PDF,
instead of being unselectable, meaningless and inefficient path data.

The result should be a PDF containing properly typeset text and graphics
(and closely matching the SVG/Inkscape design) which can be included into
*another* LaTeX document such as a paper or thesis.  The advantage of
building a PDF for inclusion instead of using svg2latex and actually
typesetting the figure as part of the main document's LaTeX job is that
the figure's design can't be broken by some change in the surrounding
document code, and the whole figure is a single file so it's effectively
an atomic unit that you can copy or rename or move without breaking
references or having to fix up file paths.  It also means that LaTeX
doesn't have to spend time re-typesetting the same figure labels every
time you build your document, although that's probably a minor advantage
unless your figure uses some very complicated typesetting in its labels.